+++ config.js
@@ -1,0 +1,8 @@
+const aliases = {
+  shared: 'src/shared',
+  posts: 'src/posts',
+}
+
+module.exports = { 
+  aliases
+}
