--- src/App/Posts/PostDetailsWrapper/PostDetailsWrapper.scss
@@ -1,1 +1,0 @@
-.postDetailsWrapper {}
