+++ src/posts/ui/PostCard/PostCard.scss
