+++ src/posts/ui/PostDetails/PostDetails.scss
