+++ src/posts/ui/PostsLayout/PostsLayout.scss
