+++ src/posts/ui/Tabs/Tabs.scss
